refactor: split db.rs into db/ module (mod, queries, tests)
CQ-1: The 1418-line db.rs is split into a module directory with
coherent boundaries:

- db/mod.rs (334 lines) — Database struct, init/open, schema
  migrations, file tracking, row types
- db/queries.rs (548 lines) — Persistence trait impl + prune helper
- db/tests.rs (555 lines) — all tests

All external imports (use crate::db::Database) are unchanged.
No logic changes — pure structural refactor.

Part of #483.
